chore: Remove `/events`
This command is now obsolete due to the onboarding survey. With the
survey, the user's choices persist even when the user leaves, so there
is no need for them to re-add their roles upon rejoin.

Please note that during rollover, event roles need to be enabled and the
survey needs to be updated with said roles accordingly.