chore: Change version schema (#20)
* fix: Strip newlines from commit short hash

* chore: Change version number to be `year.month-commit`

The semver schema for user-facing products does not make sense to a
consumer compared to a developer. Semver versioning should communicate
how big changes are so that developers are aware ahead of time before
upgrading.

Having the version schema be `year.month-commit` gives a little
perspective on when the version of the bot was released and *exactly*
what commit the bot is running on. This is great for developers, since
bisecting issues becomes a whole lot easier when issues get reported.