feat: Replace mute role assignments with Discord timeouts for mutes (#23)
* feat: Refactor member mutes to use Discord timeouts

This removes a the responsibility of handling unmutes off of Pi-Bot and
transfers it to being fully managed by Discord. Timeouts are also
tracked within Discord's mod log.

Old unmute cron code is preserved to prevent any breaking changes on
redeployment.

* chore: Add prefix to reason to identify who initiated the timeout

* feat: Add optional reason argument to `/unmute`

* fix: Add missing interaction typing

* feat: Made `/mute` reason optional

* feat: Add support for specifying unix timestamps to `/mute`

Adds an additional argument `until` to allow for unix second timestamps
to be passed `until` is mutually exclusive to `mute_length`.

* chore: Consolidate reason message to same format for `/ban` and `/kick`

* fix: Add failure message for `/mute`

* chore: Add warning logs if setting timeout fails

* fix: Fix indefinite mute to use mute role

This role is meant to be persistent and permanent. Discord timeouts are
meant to be temporary and opnly allow a max duration of 28 days at the
time of commit.

`/unmute` will attempt to remove both the timeout and the mute role from
the user if they are assigned.

* fix: Clarify timeout reasoning for spam mute