chore: remove `/confirm` and references to Unconfirmed role (#26)
* chore: Remove `/confirm`, unconfirmed role, and welcome channel

* Add default invite prefix variable

* Add unconfirmed cleanup batch kick command

To help facilitate in migrating unconfirmed pre-onboarding users to use
the new onboarding feature once they join.

* Reimplement to assume unconfirmed role has been removed

* Add doc comments

* Add long term Discord rate limit global constant

* Add bot permission check

* Add condition to ignore bots

* Add basic interval ratelimiting

* Include sending message to kicked user

Provides information to the kicked user as to why they were kicked and
how to rejoin. The kicked user is not gauranteed to get the DM; they
have the option to reject all DMs from specific servers.

* chore: Refactor implementation to be inline

Uses existing Confirm view to simpify code flow.

* feat: Add kick count to progress and final report

* fix: Remove use of `asyncio.sleep` in favor of using `asyncio.wait_for`

* fix: Use embed for DM message

Keeps consistency in formatting with other notices that Pi-Bot sends to
the user.