feat: Add basic `/censor` commands
These commands allow for interaction with automod and is mainly a
proof-of-concept.

Note that these commands require the bot to have the "Manage Server"
permission which would greatly increase the attack surface not only for
the bot, but also all staff mods and VIPs as any staff or VIP user can
add the bot's role and inherit the Manage Server permission.

It is likely best to wait and see if Discord adds a "Manage Automod"
permission for this command to become safe to use. Even if staff can
then access automod through the GUI, it can allow for insertion of
certain phrases that are not possible to add through the UI (e.g.
phrases that contain commas, since the UI uses them as delimiters
between phrases).